% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASbeautifier-package.R
\docType{package}
\name{GWASbeautifier-package}
\alias{GWASbeautifier}
\alias{GWASbeautifier-package}
\title{GWASbeautifier}
\description{
Tools to format and visualize GWAS results from GWASpoly, GAPIT, and related pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coulombevin/GWASbeautifier}
  \item Report bugs at \url{https://github.com/coulombevin/GWASbeautifier/issues}
}

}
\author{
\strong{Maintainer}: Vincent Coulombe \email{vincent.coulombe.6@ulaval.ca} (\href{https://orcid.org/0009-0000-0594-9985}{ORCID})

Authors:
\itemize{
  \item Vincent Coulombe \email{vincent.coulombe.6@ulaval.ca} (\href{https://orcid.org/0009-0000-0594-9985}{ORCID})
}

}
\keyword{internal}
