% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot beautifier}
\usage{
plot_manhattan(
  data,
  traits = NULL,
  models = NULL,
  chrom = NULL,
  chrom_color = c("#219ebc", "#8ecae6"),
  significant_color = "#cc4f89",
  multi_model_significant_color = "#fb8500",
  point_size = 1.5,
  point_alpha = 1,
  threshold_line_color = "grey50",
  threshold_line_type = 2,
  significant_markers = NULL,
  gap_size = 0.01,
  split_models = FALSE,
  lod_thicks_count = 4
)
}
\arguments{
\item{data}{'package'.thresh : Variable of class GWASpoly.thresh or GAPIT.thresh}

\item{traits}{traits vec(str) : Vector of trait names (by default, all traits plotted)}

\item{models}{vec(str) : Vector of model names (by default, all models plotted). The order of models name will be followed from top to bottom if \verb{split_models bool=TRUE}.}

\item{chrom}{vec(str) : Vector of chromosome to plot (by default, all chromosomes plotted)}

\item{chrom_color}{vec(str, str) : List of 2 string color for alternate coloration of marker.}

\item{significant_color}{str : String color for significant marker, NULL  leaves significant markers with the same color as not significant ones.}

\item{multi_model_significant_color}{str : Color used to highlight significant markers in multi-model plots.}

\item{point_size}{double : Diameter of the points.}

\item{point_alpha}{double : Points transparency.}

\item{threshold_line_color}{str : Threshold line color.}

\item{threshold_line_type}{int : Threshold line type.}

\item{significant_markers}{data.frame : Dataframe generated by the function \code{\link[=extract_markers]{extract_markers()}} containing only the significant marker to highlight with 'significant_color'. NULL value will highlight all markers according to 'significant_color' parameter.}

\item{gap_size}{double : Total size multiplier to define gap size between chromosomes.}

\item{split_models}{bool : If \code{TRUE}, split plot models into separate panels.}

\item{lod_thicks_count}{int : Number of y breaks including 0 and max value (plots will always start at 0, if \code{lod_thicks_count=2} then one thick at 0 and one at max value).}
}
\value{
ggplot2::ggplot object.
}
\description{
Publication ready Manhattan plot generator adapted from \code{\link[GWASpoly:manhattan.plot]{GWASpoly::manhattan.plot()}}, customized and optimized for data obtained with \code{\link[GWASpoly:set.threshold]{GWASpoly::set.threshold()}} or \code{\link[=get_formatted_gapit]{get_formatted_gapit()}} command.
}
\examples{
\dontrun{
plot_manhattan(data = data_with_threshold, significant_color = NULL)
}

}
\seealso{
\itemize{
\item \code{\link[=get_formatted_gapit]{get_formatted_gapit()}} for GAPIT data importation and formatting.
\item \code{\link[=extract_markers]{extract_markers()}} for significant marker extraction.
\item \code{\link[=save_manhattan]{save_manhattan()}} for pre-formatted plot saving options.
}
}
