% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_traits.R
\name{rename_traits}
\alias{rename_traits}
\title{Renaming traits function}
\usage{
rename_traits(data, old_trait_name, new_trait_name)
}
\arguments{
\item{data}{: 'package'.thresh  GWASpoly or GAPIT threshold
formated data.}

\item{old_trait_name}{: vec(str) or str   Old trait(s) name}

\item{new_trait_name}{: vec(str) or str   New trait(s) name}
}
\value{
{.code package.thresh} object with renamed traits
}
\description{
Standardize and format raw traits name into publication-ready scientific labels.
}
\examples{
\dontrun{
renamed_data_with_threshold <- rename_traits(
   data = data_with_threshold,
   old_trait_name = "vine.maturity",
   new_trait_name = "Vine Maturity"
)

renamed_data_with_threshold <- rename_traits(
   data = data_with_threshold,
   old_trait_name = c("h", "w"),
   new_trait_name = c("Height", "Weight")
)
}
}
